+++ resources/ReleaseNotes.md
@@ -11,4 +11,5 @@
 - Allow cropping of `mag.Science` with negative duration (filters from the end)
 - Allow cropping of I-ALiRT with separate filter for primary and secondary
 - Add conversion method `eventtable` in `mag.event.Event` to simplify logic in science analysis
+- Derivative of `mag.TimeSeries` variables now has missing value as first element instead of last
 - Add tests for `mag.event.Event`, `mag.event.ModeChange`, `mag.event.RangeChange` and `mag.event.RampMode`